requirements.txt
chromadb==1.0.13
groq==0.29.0
numpy==2.3.1
pandas==2.3.0
python-dotenv==1.1.1
scikit_learn==1.7.0
sentence_transformers==4.1.0
streamlit==1.46.0
